Carrito Implementado

+++ src/main/java/com/example/demo/model/Pedido.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @ToString
 public class Pedido {
 	private Integer cantidad;
-	private Integer productoId;
+	private Long productoId;
 
 
 }
